# Faster notifications and buildQueued support