--- a/src/lib/Hydra/Event.pm
+++ b/src/lib/Hydra/Event.pm
@@ -38,7 +38,7 @@ sub new_event {
     }, $self;
 }
 
-sub interested {
+sub interestedIn {
     my ($self, $plugin) = @_;
 
     return $self->{"event"}->interestedIn($plugin);