--- a/src/lib/Hydra/Plugin.pm
+++ b/src/lib/Hydra/Plugin.pm
@@ -25,29 +25,35 @@ sub instantiate {
     return @$plugins;
 }
 
-# Called when build $build has been queued.
-sub buildQueued {
-    my ($self, $build) = @_;
-}
+# To implement behaviors in response to the following events, implement
+# the function in your plugin and it will be executed by hydra-notify.
+#
+# See the tests in t/Event/*.t for arguments, and the documentation for
+# notify events for semantics.
+#
+# # Called when build $build has been queued.
+# sub buildQueued {
+#     my ($self, $build) = @_;
+# }
 
-# Called when build $build has started.
-sub buildStarted {
-    my ($self, $build) = @_;
-}
+# # Called when build $build has started.
+# sub buildStarted {
+#     my ($self, $build) = @_;
+# }
 
-# Called when build $build has finished.  If the build failed, then
-# $dependents is an array ref to a list of builds that have also
-# failed as a result (i.e. because they depend on $build or a failed
-# dependeny of $build).
-sub buildFinished {
-    my ($self, $build, $dependents) = @_;
-}
+# # Called when build $build has finished.  If the build failed, then
+# # $dependents is an array ref to a list of builds that have also
+# # failed as a result (i.e. because they depend on $build or a failed
+# # dependeny of $build).
+# sub buildFinished {
+#     my ($self, $build, $dependents) = @_;
+# }
 
-# Called when step $step has finished. The build log is stored in the
-# file $logPath (bzip2-compressed).
-sub stepFinished {
-    my ($self, $step, $logPath) = @_;
-}
+# # Called when step $step has finished. The build log is stored in the
+# # file $logPath (bzip2-compressed).
+# sub stepFinished {
+#     my ($self, $step, $logPath) = @_;
+# }
 
 # Called to determine the set of supported input types.  The plugin
 # should add these to the $inputTypes hashref, e.g. $inputTypes{'svn'}