--- a/src/lib/Hydra/TaskDispatcher.pm
+++ b/src/lib/Hydra/TaskDispatcher.pm
@@ -112,6 +112,11 @@ sub new {
         type => "counter",
         help => "Number of tasks that have been requeued after a failure."
     );
+    $prometheus->declare(
+        "notify_plugin_no_such_plugin",
+        type => "counter",
+        help => "Number of tasks that have not been processed because the plugin does not exist."
+    );
 
     my %plugins_by_name = map { ref $_ => $_ } @{$plugins};