--- a/t/Event/BuildQueued.t
+++ b/t/Event/BuildQueued.t
@@ -1,0 +1,69 @@
+use strict;
+use warnings;
+use Setup;
+use Hydra::Event;
+use Hydra::Event::BuildQueued;
+use Test2::V0;
+use Test2::Tools::Exception;
+use Test2::Tools::Mock qw(mock_obj);
+
+my $ctx = test_context();
+
+my $db = $ctx->db();
+
+my $builds = $ctx->makeAndEvaluateJobset(
+    expression => "basic.nix"
+);
+
+subtest "Parsing build_queued" => sub {
+    like(
+        dies { Hydra::Event::parse_payload("build_queued", "") },
+        qr/one argument/,
+        "empty payload"
+    );
+    like(
+        dies { Hydra::Event::parse_payload("build_queued", "abc123\tabc123") },
+        qr/only one argument/,
+        "two arguments"
+    );
+
+    like(
+        dies { Hydra::Event::parse_payload("build_queued", "abc123") },
+        qr/should be an integer/,
+        "not an integer"
+    );
+    is(
+        Hydra::Event::parse_payload("build_queued", "19"),
+        Hydra::Event::BuildQueued->new(19),
+        "Valid parse"
+    );
+};
+
+subtest "load" => sub {
+    my $build = $builds->{"empty_dir"};
+
+    my $event = Hydra::Event::BuildQueued->new($build->id);
+
+    $event->load($db);
+
+    is($event->{"build"}->id, $build->id, "The build record matches.");
+
+    # Create a fake "plugin" with a buildQueued sub, the sub sets this
+    # global passedBuild variable.
+    my $passedBuild;
+    my $plugin = {};
+    my $mock = mock_obj $plugin => (
+        add => [
+            "buildQueued" => sub {
+                my ($self, $build) = @_;
+                $passedBuild = $build;
+            }
+        ]
+    );
+
+    $event->execute($db, $plugin);
+
+    is($passedBuild->id, $build->id, "The plugin's buildQueued hook is called with the proper build");
+};
+
+done_testing;