@grahamc 2021-12-21
I've had this deployed on Hydra for a bit. We recently had an evaluation register ~50k new builds, and hydra-notify was able to process all of them within less than 1 sample period (ie: 15s):

![image](https://user-images.githubusercontent.com/76716/146998097-5c7cd379-eae5-45a5-baca-c35060908f69.png)