@grahamc 2021-12-21
Another bit of data to look at: the CompressLogs is _intended_ to run at the end of every step being finished. You can see that happening here:

![image](https://user-images.githubusercontent.com/76716/147000260-2fac8951-43aa-40e0-9cbd-1885ba5e82d7.png)

Compare to GithubPulls which used to run on every step finishing too, but no longer does so since it has nothing to do:

![image](https://user-images.githubusercontent.com/76716/147000366-321b117a-6b75-4e7f-93ad-1d44c014ec97.png)

Note in both of these the Y axis is the total # of invocations.