@grahamc 2021-12-22
I added a new metric to count how much time is spent processing events in total, see: https://monitoring.nixos.org/prometheus/graph?g0.expr=rate(notify_event_runtime_sum[10m])&g0.tab=0&g0.stacked=0&g0.show_exemplars=0&g0.range_input=1h this will let us know how this PR is affecting the notify daemon.