@grahamc 2021-12-23
I think the data is pretty compelling and that we're not drastically harming the performance of hydra-notify by turning build-queued on again.

![image](https://user-images.githubusercontent.com/76716/147252947-c271a42e-a595-4eea-920c-6a11d6b4e6fd.png)